requirements.txt
fastapi>=0.95.0
uvicorn>=0.20.0
httpx==0.28.1
python-multipart==0.0.20

motor==3.3.1
pymongo==4.6.3

bcrypt
passlib
python-jose

sqlalchemy
psycopg2-binary
asyncpg

python-dotenv
